...G...... [ZBXNEXT-6210] return non percentage based proc.mem statistics as uint64.

+++ src/go/plugins/proc/proc_linux.go
@@ -585,7 +585,7 @@ func (p *PluginExport) Export(key string, params []string, ctx plugin.ContextPro
 	}
 
 	if memtype != "pmem" {
-		return int(memSize), nil
+		return uint64(memSize), nil
 	}
 	return memSize, nil
 }
